requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
--extra-index-url https://www.piwheels.hostedpi.com/simple

aniso8601==2.0.0          # via flask-restplus
attrs==17.4.0
click==6.7                # via flask
flask-restplus==0.10.1
flask==1.0.2
gunicorn==19.9.0
itsdangerous==0.24        # via flask
jinja2==2.10              # via flask
jsonschema==2.6.0         # via flask-restplus
markupsafe==1.0           # via jinja2
paho-mqtt==1.4.0
pytz==2017.3              # via flask-restplus
rpi-rf==0.9.7
rpi.gpio==0.6.5           # via rpi-rf
schema==0.6.7
six==1.11.0               # via flask-restplus
werkzeug==0.14.1          # via flask
